Main_code_web_01
So i started to learn html and css today, and i want to put my first lessons right into move and do something useful so i decided to begin working on our webpage at least template of it. We don't have to do it by ourselves, we can do it trough wordpress or other alrdy signed and trusted webhosters but its good to know hot to edit some stuff. Feel free to add whats on your mind, if you have some comment and want to show it in the file use please <!-- yourtexthere normaly :) --> . If you have any other ideas please write a comment on it.